// src/providers/passage/dto/create-passage-user.dto.ts
export class CreatePassageUserDto {
  user: {
    email: string;
    password: string;
    password_confirmation: string;
    gsm: string;
    first_name: string;
    last_name: string;
    job_position_id: string | null;
    activation_date: string;
    citizenship_number: string;
    department_id: string | null;
    username: string;
    about: string;
    is_active: boolean;
    client_id: string;
    gender: string;
    birth_date: string;
    branch_id: string | null;
    job_departure_date: string | null;
    user_type: string;
    consumed_timestep: string | null;
    sub_company_id: string | null;
    nationality: string;
    marital_status: string;
    spouses_employment_status: string;
    disability_level: string;
    number_of_children: string;
    educational_status: string;
    graduation_level: string;
    graduation_school: string;
    bank_name: string;
    account_type: string;
    account_number: string;
    iban: string;
    emergency_contact_person: string;
    emergency_person_proximity_degree: string;
    emergency_contact_person_1: string;
    emergency_person_proximity_degree_1: string;
    emergency_contact_phone: string;
    other_gsm: string | null;
    organization_unit_id: string | null;
    title_id: string | null;
    user_group_id: string | null;
    expired_date: string | null;
    reason_for_leave_id: string | null;
    shift_id: string | null;
    faculty_name: string;
    university_department: string;
    employee_type: string;
    supervisor_registration_number: string;
    branch_code: string;
    department_name: string;
    department_code: string;
    job_position_name: string;
    job_position_code: string;
    title_name: string;
    title_code: string;
    sub_company_name: string;
    sub_company_code: string;
    employee_id: string;
    group_name: string;
    group_code: string;
    organization_unit_name: string;
    organization_unit_code: string;
  };
}